<center> tags replaced
The html validator mentioned the <center> tags were basically obsolete and to use css instead. I deleted all of the <center> tags and changed the areas with text to style="text-align: center". I added divs to the youtube video and the overwatch image and added the text align center within it. The website itself didn't change however the code did.