Updated Website + CSS (Fixed two major issues)
I solved 2 main issues I was having. Spacing apart the zenyatta and lucio images evenly and getting rid of the extra space I had on the right side of my website. I deleted the div class row from html and the css code with it. This solved the extra space on the right side issue. Then I changed the float to right and fixed the margins and width so that the images would be spaced evenly which solved the second issue.